refactor(air): generalize lookup validation; split into focused files
Pulls the Miden-specific debug validation surface into the generic
`air/src/lookup/debug/validation/` tree so only the Miden AIR layout
and degree budget stay Miden-side.

Generalized:
- `check_symbolic_degrees` takes `AirLayout` + `degree_budget` as
  parameters (no more `ProcessorAir`/`TRACE_WIDTH`/`DEGREE_BUDGET = 9`
  references inside).
- `validate` + `ValidationReport` moved from
  `constraints/lookup/debug/validation/mod.rs` into the generic tree;
  takes the same `AirLayout` + `degree_budget` parameters.

Miden glue: the sole consumer is `MidenLookupAir`'s test module, which
now inlines `DEGREE_BUDGET = 9` + a `miden_air_layout()` helper and calls
the generic `check_symbolic_degrees` directly. The
`constraints/lookup/debug/` directory is deleted entirely.

Split `lookup/debug/validation/` into focused files:
- `records.rs` — inventory data types (`DebugStructure`, `GroupRecord`,
  …) + their view/display impls
- `builder.rs` — `DebugStructureBuilder` (unchanged)
- `inspect.rs` — walker entry points (`inspect_structure`,
  `collect_inventory`, `check_encoding_equivalence`,
  `check_challenge_scoping`)
- `symbolic.rs` — `check_symbolic_degrees` (generalized)
- `structure.rs` — `StructureReport` + `validate_structure_only`
- `validate.rs` — `ValidationReport` + composed `validate`
- `mod.rs` — layered-check table doc + re-exports

Verified: `make check`, `make lint`,
`cargo test -p miden-air --lib` (50 pass),
`cargo test -p miden-processor --lib trace::tests::lookup` (3 pass).
