+++ air/src/lookup/debug/mod.rs
@@ -4,12 +4,12 @@
 //!
 //! | Module | Regime |
 //! |--------|--------|
-//! | [`validation`] | AIR self-checks — run against the `LookupAir` itself, no execution trace needed. Bundles encoding equivalence, scope, and constants consistency via [`validation::validate_structure_only`]. |
+//! | [`validation`] | AIR self-checks — run against the `LookupAir` itself, no execution trace needed. Bundles encoding equivalence, scope, constants consistency, and the symbolic degree-budget pass via [`validation::validate`]. |
 //! | [`trace`] | Concrete-trace debugging — balance accumulator + per-column `(U, V)` oracle folds + mutex checks over a real main trace. |
 //!
-//! The symbolic degree-budget pass (`check_symbolic_degrees`) and the composed
-//! `validate` entry point live Miden-side at
-//! [`crate::constraints::lookup::debug::validation`].
+//! `validation::validate` takes a caller-supplied `AirLayout` and degree budget, so
+//! Miden-specific constants stay Miden-side in
+//! [`crate::constraints::lookup::validation`].
 
 pub mod trace;
 pub mod validation;
@@ -28,8 +28,9 @@ pub use trace::{
     collect_column_oracle_folds,
 };
 pub use validation::{
-    ColumnRecord, DebugStructure, DebugStructureBuilder, EncodingMode, GroupMismatch, GroupRecord,
-    InteractionRecord, Inventory, MultSign, NumColumnsCheck, ScopeReport, StructureReport,
-    check_challenge_scoping, check_encoding_equivalence, collect_inventory, inspect_structure,
-    validate_structure_only,
+    ColumnRecord, DebugStructure, DebugStructureBuilder, DegreeMismatch, DegreeReport,
+    EncodingMode, GroupMismatch, GroupRecord, InteractionRecord, Inventory, MultSign,
+    NumColumnsCheck, ScopeReport, StructureReport, ValidationReport, check_challenge_scoping,
+    check_encoding_equivalence, check_symbolic_degrees, collect_inventory, inspect_structure,
+    validate, validate_structure_only,
 };
